@@ -6,9 +6,9 @@
   "license": "UNLICENSED",
   "scripts": {
     "build": "tsc",
-    "start": "node -r ts-node/register -r tsconfig-paths/register -r elastic-apm-node/start build/main.js",
-    "dev": "ts-node -r tsconfig-paths/register -r elastic-apm-node/start src/main.ts",
-    "dev:watch": "ts-node-dev --respawn --transpileOnly -r tsconfig-paths/register -r elastic-apm-node/start src/main.ts",
+    "start": "node -r ts-node/register -r tsconfig-paths/register build/main.js",
+    "dev": "ts-node -r tsconfig-paths/register src/main.ts",
+    "dev:watch": "ts-node-dev --respawn --transpileOnly -r tsconfig-paths/register src/main.ts",
     "type": "tsc --noEmit",
     "type:watch": "tsc --noEmit --watch",
     "format": "prettier --write \"src/**/*.ts\"",
@@ -21,61 +21,62 @@
     "test:e2e:cov": "jest --config jest.e2e-config.json --coverage"
   },
   "dependencies": {
-    "@azure/storage-blob": "^12.1.0",
-    "@nestjs/common": "6.10.14",
-    "@nestjs/core": "6.10.14",
-    "@nestjs/platform-express": "6.10.14",
-    "@nestjs/swagger": "4.1.2",
+    "@azure/storage-blob": "12.1.1",
+    "@nestjs/common": "7.0.5",
+    "@nestjs/core": "7.0.5",
+    "@nestjs/platform-express": "7.0.5",
+    "@nestjs/swagger": "4.4.0",
     "@panva/jose": "1.9.3",
+    "applicationinsights": "^1.7.3",
     "axios": "0.19.2",
     "base64url": "3.0.1",
     "bson-objectid": "1.3.0",
     "class-transformer": "0.2.3",
-    "class-validator": "0.11.0",
+    "class-validator": "0.11.1",
     "compression": "1.7.4",
-    "cookie-parser": "1.4.4",
-    "elastic-apm-node": "3.3.0",
-    "envalid": "6.0.0",
-    "helmet": "3.21.2",
+    "cookie-parser": "1.4.5",
+    "elastic-apm-node": "3.5.0",
+    "envalid": "6.0.1",
+    "helmet": "3.21.3",
     "moment": "2.24.0",
-    "nodemailer": "6.4.2",
-    "nunjucks": "3.2.0",
-    "pg": "7.14.0",
+    "nodemailer": "6.4.6",
+    "nunjucks": "3.2.1",
+    "pg": "7.18.2",
     "reflect-metadata": "0.1.13",
-    "rxjs": "6.5.3",
-    "swagger-ui-express": "4.1.2",
-    "typeorm": "0.2.22"
+    "rxjs": "6.5.4",
+    "swagger-ui-express": "4.1.4",
+    "typeorm": "0.2.24"
   },
   "devDependencies": {
-    "@nestjs/testing": "6.10.14",
-    "@types/compression": "1.0.1",
+    "@nestjs/testing": "7.0.5",
+    "@types/compression": "1.7.0",
     "@types/cookie-parser": "1.4.2",
-    "@types/express": "4.17.2",
-    "@types/faker": "4.1.8",
+    "@types/express": "4.17.3",
+    "@types/faker": "4.1.10",
     "@types/helmet": "0.0.45",
-    "@types/jest": "25.1.1",
+    "@types/jest": "25.1.4",
     "@types/multer": "^1.4.2",
-    "@types/node": "12.12.17",
+    "@types/node": "13.9.3",
     "@types/nodemailer": "6.4.0",
     "@types/nunjucks": "3.1.3",
     "@types/supertest": "2.0.8",
-    "@typescript-eslint/eslint-plugin": "2.11.0",
-    "@typescript-eslint/parser": "2.11.0",
-    "eslint": "6.7.2",
-    "eslint-config-prettier": "6.7.0",
-    "eslint-plugin-jest": "23.1.1",
+    "@typescript-eslint/eslint-plugin": "2.24.0",
+    "@typescript-eslint/parser": "2.24.0",
+    "eslint": "6.8.0",
+    "eslint-config-prettier": "6.10.1",
+    "eslint-plugin-jest": "23.8.2",
     "eslint-plugin-promise": "4.2.1",
     "eslint-plugin-security": "1.4.0",
     "faker": "4.1.0",
-    "husky": "3.1.0",
+    "husky": "4.2.3",
     "jest": "25.1.0",
-    "lint-staged": "9.5.0",
-    "prettier": "1.19.1",
+    "lint-staged": "10.0.8",
+    "prettier": "2.0.1",
     "supertest": "4.0.2",
-    "ts-jest": "25.2.0",
-    "ts-node": "8.5.4",
+    "ts-jest": "25.2.1",
+    "ts-node": "8.8.1",
     "ts-node-dev": "1.0.0-pre.44",
     "tsconfig-paths": "3.9.0",
-    "typescript": "3.7.3"
+    "typescript": "3.8.3"
   }
 }
