@@ -46,9 +46,10 @@ export class MockConfigService extends Config {
       EMAIL_CHANGE_TOKEN_LIFETIME_MIN: 0,
       SESSION_NAME: '',
       SESSION_MAX_AGE_MIN: 0,
-      ELASTIC_APM_SERVICE_NAME: '',
-      ELASTIC_APM_SECRET_TOKEN: '',
-      ELASTIC_APM_SERVER_URL: '',
+      // ELASTIC_APM_SERVICE_NAME: '',
+      // ELASTIC_APM_SECRET_TOKEN: '',
+      // ELASTIC_APM_SERVER_URL: '',
+      AZURE_MONITOR_INSTRUMENTATION_KEY: '',
       AZURE_BLOB_STORAGE_CONNECTION_STRING: '',
     };
   }
