@@ -1,5 +1,5 @@
 import { Injectable, NotImplementedException } from '@nestjs/common';
-import { Config, InjectConfig } from 'config/application/Config';
+import { Config } from 'config/application/Config';
 import { EmailManager } from 'email/EmailManager';
 import axios from 'axios';
 
@@ -11,7 +11,7 @@ export class SendgridEmailService implements EmailManager {
   private readonly sendgridApiKey: string;
   private readonly sendgridUrl: string;
 
-  public constructor(@InjectConfig() _config: Config) {
+  public constructor(_config: Config) {
     // this.sendgridApiKey = config.get('SENDGRID_API_KEY');
     // this.sendgridUrl = config.get('SENDGRID_URL');
     this.sendgridApiKey = '';
