@@ -51,7 +51,7 @@ export class DomainEventHandlerManagerService
     for (const module of this.modulesContainer.values()) {
       for (const instanceWrapper of module.providers.values()) {
         const { instance } = instanceWrapper;
-        if (!instance) {
+        if (typeof instance !== 'object' || !instance) {
           continue;
         }
         const metadataItems:
