@@ -1,13 +1,6 @@
-import { Inject } from '@nestjs/common';
 import { Contribution } from 'project/domain/value-objects/Contribution';
 import { PeerReviewCollection } from 'project/domain/PeerReviewCollection';
 import { Id } from 'common/domain/value-objects/Id';
-
-export const CONTRIBUTIONS_COMPUTER = Symbol('CONTRIBUTIONS_COMPUTER');
-
-export function InjectContributionsComputer(): ParameterDecorator {
-  return Inject(CONTRIBUTIONS_COMPUTER);
-}
 
 /**
  *
@@ -19,9 +12,9 @@ export interface Contributions {
 /**
  * Contributions Computer
  */
-export interface ContributionsComputer {
+export abstract class ContributionsComputer {
   /**
    * Computes the relative contributions.
    */
-  compute(peerReviews: PeerReviewCollection): Contributions;
+  public abstract compute(peerReviews: PeerReviewCollection): Contributions;
 }
