@@ -5,12 +5,17 @@ import { User } from 'user/domain/User';
 import { ModelCollection } from 'common/domain/ModelCollection';
 import { Contributions } from 'project/domain/ContributionsComputer';
 import { RoleNoUserAssignedException } from 'project/domain/exceptions/RoleNoUserAssignedException';
+import { SingleAssignmentPerUserViolationException } from 'project/domain/exceptions/SingleAssignmentPerUserViolationException';
 
 export class RoleCollection extends ModelCollection<Role> {
   public static empty(): RoleCollection {
     return new RoleCollection([]);
   }
 
+  /**
+   * Find role by assignee.
+   * @param assigneeOrAssigneeId
+   */
   public findByAssignee(assigneeOrAssigneeId: User | Id): Role {
     for (const role of this) {
       if (role.isAssignedToUser(assigneeOrAssigneeId)) {
@@ -20,11 +25,11 @@ export class RoleCollection extends ModelCollection<Role> {
     throw new RoleNotFoundException();
   }
 
-  public excluding(roleToExclude: Role): ReadonlyArray<Role> {
+  public excluding(roleToExclude: Role): Iterable<Role> {
     if (!this.exists(roleToExclude.id)) {
       throw new RoleNotFoundException();
     }
-    return this.toArray().filter(role => !role.equals(roleToExclude));
+    return this.filter(role => !role.equals(roleToExclude));
   }
 
   public applyContributions(contributions: Contributions): void {
@@ -34,16 +39,32 @@ export class RoleCollection extends ModelCollection<Role> {
   }
 
   public anyAssignedToUser(userOrUserId: User | Id): boolean {
-    return this.toArray().some(role => role.isAssignedToUser(userOrUserId));
+    return this.any(role => role.isAssignedToUser(userOrUserId));
+  }
+
+  public assertSingleAssignmentPerUser(): void {
+    const seenUsers: Id[] = [];
+    for (const role of this) {
+      if (role.assigneeId) {
+        for (const seenUser of seenUsers) {
+          if (role.isAssignedToUser(seenUser)) {
+            throw new SingleAssignmentPerUserViolationException();
+          }
+        }
+        seenUsers.push(role.assigneeId);
+      }
+    }
   }
 
   public assertAllAreAssigned(): void {
-    if (!this.toArray().every(role => role.isAssigned())) {
-      throw new RoleNoUserAssignedException();
+    for (const role of this) {
+      if (!role.isAssigned()) {
+        throw new RoleNoUserAssignedException();
+      }
     }
   }
 
   public allHaveSubmittedPeerReviews(): boolean {
-    return this.toArray().every(role => role.hasSubmittedPeerReviews.value);
+    return this.all(role => role.hasSubmittedPeerReviews.value);
   }
 }
